Phase-3/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
anthropic>=0.18.0
openai>=1.12.0
mcp>=1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.9
alembic==1.12.1
sqlalchemy==2.0.23
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
